refactor: move hasSingleTypeFilter to filter helpers, use $derived for table columns
- Move hasSingleTypeFilter from table/options.svelte.ts to filters/helpers.svelte.ts
  where it belongs as a filter concern
- Replace createTableState + $effect with $derived(configuration.columns) in
  getSharedTableOptions — eliminates the state-to-state sync anti-pattern and
  naturally tracks the reactive column getter without any side effects
- Update imports in events, stacks, stream pages accordingly
