+++ src/Exceptionless.Web/ClientApp/src/lib/features/events/components/filters/helpers.svelte.ts
@@ -203,6 +203,17 @@ export function shouldRefreshPersistentEventChanged(
     return true;
 }
 
+const TYPE_FILTER_REGEX = /\btype:(\w+)\b/g;
+
+export function hasSingleTypeFilter(filter: null | string | undefined): boolean {
+    if (!filter) {
+        return false;
+    }
+
+    const matches = filter.match(TYPE_FILTER_REGEX);
+    return matches?.length === 1;
+}
+
 export function toFilter(filters: IFilter[]): string {
     return filters
         .map((f) => f.toFilter())
