+++ src/Exceptionless.Web/ClientApp/src/lib/features/shared/table.svelte.ts
@@ -68,11 +68,7 @@ export function getSharedTableOptions<TData extends RowData, TPaginationStrategy
     const isMemoryPaging = $derived(configuration.paginationStrategy === 'memory');
 
     const [pageCount, setPageCount] = createTableState(0);
-    const [columns, setColumns] = createTableState<ColumnDef<StockFeatures, TData, unknown>[]>(configuration.columns);
-
-    $effect(() => {
-        setColumns(configuration.columns);
-    });
+    const columns = $derived(configuration.columns);
 
     // Use the persistKey if provided, otherwise default to events-column-visibility
     const visibilityKey = configuration.columnPersistenceKey ? `${configuration.columnPersistenceKey}-column-visibility` : 'events-column-visibility';
@@ -84,9 +80,9 @@ export function getSharedTableOptions<TData extends RowData, TPaginationStrategy
 
     const orderKey = configuration.columnPersistenceKey ? `${configuration.columnPersistenceKey}-column-order` : 'events-column-order';
     const [persistedColumnOrder, setPersistedColumnOrder] = createPersistedTableState(orderKey, configuration.defaultColumnOrder ?? <ColumnOrderState>[]);
-    const columnOrder = () => sanitizeColumnOrder(persistedColumnOrder(), columns());
+    const columnOrder = () => sanitizeColumnOrder(persistedColumnOrder(), columns);
     const setColumnOrder = (updaterOrValue: Updater<ColumnOrderState>) => {
-        setPersistedColumnOrder(updaterOrValue instanceof Function ? updaterOrValue(resolveColumnOrder(columnOrder(), columns())) : updaterOrValue);
+        setPersistedColumnOrder(updaterOrValue instanceof Function ? updaterOrValue(resolveColumnOrder(columnOrder(), columns)) : updaterOrValue);
     };
 
     // Initialize pagination state from parameters
@@ -230,10 +226,7 @@ export function getSharedTableOptions<TData extends RowData, TPaginationStrategy
         _features: stockFeatures,
         _rowModels: { coreRowModel: createCoreRowModel<StockFeatures, TData>() },
         get columns() {
-            return columns();
-        },
-        set columns(value) {
-            setColumns(value);
+            return columns;
         },
         get data() {
             return data();
