+++ src/Exceptionless.Web/ClientApp/src/routes/(app)/events/+page.svelte
@@ -21,6 +21,7 @@
         filterChanged,
         filterRemoved,
         getFiltersFromCache,
+        hasSingleTypeFilter,
         serializeFilters,
         shouldRefreshPersistentEventChanged,
         toFilter,
@@ -29,7 +30,7 @@
     import OrganizationDefaultsFacetedFilterBuilder from '$features/events/components/filters/organization-defaults-faceted-filter-builder.svelte';
     import EventsBulkActionsDropdownMenu from '$features/events/components/table/events-bulk-actions-dropdown-menu.svelte';
     import EventsDataTable from '$features/events/components/table/events-data-table.svelte';
-    import { defaultEventColumnVisibility, getColumns, hasSingleTypeFilter } from '$features/events/components/table/options.svelte';
+    import { defaultEventColumnVisibility, getColumns } from '$features/events/components/table/options.svelte';
     import { organization } from '$features/organizations/context.svelte';
     import SavedViewPicker from '$features/saved-views/components/saved-view-picker.svelte';
     import { useSavedViews } from '$features/saved-views/use-saved-views.svelte';
