+++ src/Exceptionless.Web/ClientApp/src/routes/(app)/stacks/+page.svelte
@@ -19,13 +19,14 @@
         filterChanged,
         filterRemoved,
         getFiltersFromCache,
+        hasSingleTypeFilter,
         serializeFilters,
         toFilter,
         updateFilterCache
     } from '$features/events/components/filters/helpers.svelte';
     import OrganizationDefaultsFacetedFilterBuilder from '$features/events/components/filters/organization-defaults-faceted-filter-builder.svelte';
     import EventsDataTable from '$features/events/components/table/events-data-table.svelte';
-    import { getColumns, hasSingleTypeFilter } from '$features/events/components/table/options.svelte';
+    import { getColumns } from '$features/events/components/table/options.svelte';
     import { organization } from '$features/organizations/context.svelte';
     import SavedViewPicker from '$features/saved-views/components/saved-view-picker.svelte';
     import { useSavedViews } from '$features/saved-views/use-saved-views.svelte';
